% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaregressionintervals.R
\name{interpret_anova_regression_intervals}
\alias{interpret_anova_regression_intervals}
\title{Confidence Interval and Prediction Interval from Two-way ANOVA Model}
\usage{
interpret_anova_regression_intervals(formula, data, newdata, alpha = 0.05)
}
\arguments{
\item{formula}{A formula like y ~ group1*group2 (with interaction) or y ~ group1 + group2 (without interaction). (One group variable must be numeric).}

\item{data}{A data frame.}

\item{alpha}{Significance level (default 0.05).}

\item{x_value}{A value of x-variable (One variable must be numeric).}
}
\value{
Printed List of CI, Interpretation of CI, PI, and Interpretation of PI.
}
\description{
When given a specific values for each factor, function generates confidence interval and prediction for that value based on a two-way ANOVA model
}
