% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayCIPI.R
\name{interpret_one_anova_intervals}
\alias{interpret_one_anova_intervals}
\title{Confidence Interval and Prediction Interval from One-way ANOVA Model}
\usage{
interpret_one_anova_intervals(formula, data, x_value, alpha = 0.05)
}
\arguments{
\item{formula}{A formula like y ~ group (group variable must be numeric).}

\item{data}{A data frame.}

\item{x_value}{A value of x-variable (must be numeric).}

\item{alpha}{Significance level (default 0.05).}
}
\value{
Printed List of CI, Interpretation of CI, PI, and Interpretation of PI.
}
\description{
When given a specific x value, function generates confidence interval and prediction for that value basis on a one-way ANOVA Model
}
