% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayregressionplot.R
\name{plot_anova_regression_ci_pi}
\alias{plot_anova_regression_ci_pi}
\title{Regression Plot One-way ANOVA with CI and PI}
\usage{
plot_anova_regression_ci_pi(
  formula,
  data,
  alpha = 0.05,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{formula}{A formula like y ~ group (group variable must be numeric).}

\item{data}{A data frame.}

\item{alpha}{Significance level (default 0.05).}
}
\value{
A scatter plot with regression line, confidence interval band, and prediction interval band
}
\description{
Fits a linear regression based on ANOVA model and returns scatter plot with regression line, CI band, and PI band.
}
