% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayR2.R
\name{report_oneway_anova_r2}
\alias{report_oneway_anova_r2}
\title{One_way ANOVA extract and Interpret R-Squared}
\usage{
report_oneway_anova_r2(formula, data)
}
\arguments{
\item{formula}{A formula like y ~ group.}

\item{data}{A data frame.}
}
\value{
A report containing response variable, grouping variable, R_squared, and interpretation.
}
\description{
Extracts R-squared value and prints a sentences detailing \% of variation explaining anova model and \% of variation due to variability within groups
}
