% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overpval1.R
\name{report_overall_pval_anova}
\alias{report_overall_pval_anova}
\title{Report Anova P-value}
\usage{
report_overall_pval_anova(formula, data, alpha = 0.05)
}
\arguments{
\item{formula}{A formula like y ~ group.}

\item{data}{A data frame.}

\item{alpha}{Significance level (default 0.05).}
}
\value{
A caption including model, p-value (<0.0001 printed if value is less than 0.0001), significance level (alpha), and if the effect of the variable is significant at the chosen significance level.
}
\description{
Reports and determines the significance
}
