% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twowaypvalues.R
\name{report_pvals_anova}
\alias{report_pvals_anova}
\title{Report Anova P-values for each factor and interaction}
\usage{
report_pvals_anova(formula, data, alpha = 0.05)
}
\arguments{
\item{formula}{A formula like y ~ group1*group2 (with interaction) or y ~ group1 + group2 (without interaction).}

\item{data}{A data frame.}

\item{alpha}{Significance level (default 0.05).}
}
\value{
A list of p-values and interpretation for each factor (also p-value and interpretation for interaction if model includes interaction)
}
\description{
Reports and determines the significance.
}
