% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twowayflextable.R
\name{twoway_anova_flextable}
\alias{twoway_anova_flextable}
\title{Two-Way ANOVA Flextable
Fits a two-way ANOVA model and returns three flextables, ANOVA table, model summary, and parameter estimates, for formulas that have 2 categorical or 1 categorical/1 numeric.}
\usage{
twoway_anova_flextable(formula, data, alpha = 0.05)
}
\arguments{
\item{data}{A data frame.}

\item{alpha}{Significance level (default 0.05).}
}
\value{
A list of flextable objects.
}
\description{
@param formula A formula like y ~ group1*group2 (with interaction) or y ~ group1 + group2 (without interaction).
}
