% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twowayplots.R
\name{twoway_anova_plots}
\alias{twoway_anova_plots}
\title{Plots for Two-Way Anova}
\usage{
twoway_anova_plots(data, response, factor1, factor2)
}
\arguments{
\item{data}{A data frame.}

\item{response}{response variable}

\item{factor1}{factor/grouping variable 2}
}
\value{
4 Plots, Interaction Plot, Main Effect Factor1, Main Effect Factor2, Grouped Box Plot
}
\description{
Generates 4 Plots, Interaction Plot, Main Effect Factor1, Main Effect Factor2, Grouped Box Plot
}
